+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod cli;
 mod diff;
 mod input;
 mod load;
+mod profile;
 mod syntax;
 mod transform;
 mod viewer;